Route approximation on NetworkRouteSelector

+++ OsmAnd-java/src/main/java/net/osmand/util/MapUtils.java
@@ -742,7 +742,7 @@ public class MapUtils {
 	}
 
 	public static double getSqrtDistance(int startX, int startY, int endX, int endY) {
-		return Math.sqrt((endX - startX) * (endX - startX) + (endY - startY) * (endY - startY));
+		return Math.sqrt((long) (endX - startX) * (endX - startX) + (long) (endY - startY) * (endY - startY));
 	}
 
 	public static double getSqrtDistance(float startX, float startY, float endX, float endY) {
